clearer structure to ease debugging
