+++ .gitignore
@@ -8,3 +8,4 @@ saved_model/
 latest_training_predictions/
 datasets/
 correct.engine
+build
